Move WalletService into Bisq 2 specific code
