+++ apps/rest-api-app/src/main/java/bisq/rest_api/RestApiApplicationService.java
@@ -37,8 +37,8 @@ import bisq.settings.SettingsService;
 import bisq.support.SupportService;
 import bisq.trade.TradeService;
 import bisq.user.UserService;
-import bisq.wallets.core.BitcoinWalletSelection;
 import bisq.wallets.core.WalletService;
+import bisq.wallets.core.BitcoinWalletSelection;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
