+++ bisq-easy/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
     implementation("network:network")
     implementation("network:network-identity")
     implementation("bitcoind:core")
+    implementation("wallets:wallet")
     // implementation("wallets:electrum")
     // implementation("wallets:bitcoind")
